backend/requirements.txt
# Requirements for Credit Decision Coach Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
httpx==0.26.0
supabase==2.3.4
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
